Movement Basics
I implemented the new input system into the project. We are able to walk, jump, and crouch. Tried to start fluttering but could not figure it out at this time. Additionally attached the following camera to the player.

If the player moves to the side of a platform, they will stick to it, which is not intended. Will likely need to shrink the ceiling and grounded hit boxes.